refactor: changed the file layout, but the module structure remains as is.
